Transactions service (#70)
* add(transactions): init module

* add: bases of the transactions service

* add: database

* add: transaction service

* fix: changed env variable

* fix: merge

* fix: .env

* fix: transaction service

* add: newline

* fix: newline

+++ funds/domain/funds.go
@@ -33,7 +33,7 @@ func (id FundsID) String() string {
 type Funds struct {
 	ID          FundsID   `json:"id"`
 	UserId      string    `json:"user_id"`
-	Balance     int       `json:"balance"`
+	Balance     float64   `json:"balance"`
 	LastUpdated time.Time `json:"last_updated"`
 }
